+++ .assets/provision/setup_profiles_allusers.ps1
@@ -8,31 +8,34 @@ sudo .assets/provision/setup_profiles_allusers.ps1
 $ErrorActionPreference = 'SilentlyContinue'
 $WarningPreference = 'Ignore'
 
-# path varaibles
-$PROFILE_PATH = [IO.Path]::GetDirectoryName($PROFILE.AllUsersAllHosts)
+# path variables
 $SCRIPTS_PATH = '/usr/local/share/powershell/Scripts'
+# determine folder with config files
+$CFG_PATH = (Test-Path /tmp/config -PathType Container) ? '/tmp/config' : '.assets/config'
 
 # *Copy global profiles
-if (Test-Path /tmp/config/pwsh_cfg -PathType Container) {
+if (Test-Path $CFG_PATH/pwsh_cfg -PathType Container) {
     # PowerShell profile
-    Move-Item /tmp/config/pwsh_cfg/profile.ps1 -Destination $PROFILE_PATH
+    Copy-Item $CFG_PATH/pwsh_cfg/profile.ps1 -Destination $PROFILE.AllUsersAllHosts
     # PowerShell functions
     New-Item $SCRIPTS_PATH -ItemType Directory | Out-Null
-    Move-Item /tmp/config/pwsh_cfg/ps_aliases_common.ps1 -Destination $SCRIPTS_PATH
+    Copy-Item $CFG_PATH/pwsh_cfg/ps_aliases_common.ps1 -Destination $SCRIPTS_PATH
     # git functions
     if (Test-Path /usr/bin/git -PathType Leaf) {
-        Move-Item /tmp/config/pwsh_cfg/ps_aliases_git.ps1 -Destination $SCRIPTS_PATH
+        Copy-Item $CFG_PATH/pwsh_cfg/ps_aliases_git.ps1 -Destination $SCRIPTS_PATH
     }
     # kubectl functions
     if (Test-Path /usr/bin/kubectl -PathType Leaf) {
-        Move-Item /tmp/config/pwsh_cfg/ps_aliases_kubectl.ps1 -Destination $SCRIPTS_PATH
+        Copy-Item $CFG_PATH/pwsh_cfg/ps_aliases_kubectl.ps1 -Destination $SCRIPTS_PATH
     }
-    # clean config folder
-    Remove-Item -Force -Recurse /tmp/config/pwsh_cfg
+    if (Test-Path /tmp/config/pwsh_cfg -PathType Container) {
+        # clean config folder
+        Remove-Item -Force -Recurse /tmp/config/pwsh_cfg
+    }
 }
 
 # *PowerShell profile
-if (-not ((Get-Module PowerShellGet -ListAvailable -ErrorAction SilentlyContinue).Version.Major -ge 3)) {
+while (-not ((Get-Module PowerShellGet -ListAvailable).Version.Major -ge 3)) {
     Write-Host 'installing PowerShellGet...'
     Install-Module PowerShellGet -AllowPrerelease -Scope AllUsers -Force
 }
@@ -40,7 +43,7 @@ if (-not (Get-PSResourceRepository -Name PSGallery).Trusted) {
     Write-Host 'setting PSGallery trusted...'
     Set-PSResourceRepository -Name PSGallery -Trusted
 }
-if (-not (Get-Module posh-git -ListAvailable)) {
+while (-not (Get-Module posh-git -ListAvailable)) {
     Write-Host 'installing posh-git...'
     Install-PSResource -Name posh-git -Scope AllUsers
 }
